do adj and changes
