+++ src/js/store.js
@@ -1,0 +1,3 @@
+import {writable} from "svelte/store";
+
+export let score = writable(1);
